issue: 4963823 Split tuning report docs
Separate the monolithic xlio_tuning_report_reference.md into two
purpose-built documents:

- Add docs/llm/xlio_tuning_report_guide.md: self-contained LLM
  guide with the full analysis workflow, troubleshooting rules for
  all 21 WARNINGs, 9 scenario snippets, config knob quick reference,
  field-to-config mapping, report comparison guidance, and known
  limitations.
- Refactor docs/xlio_tuning_report_reference.md into a compact
  human reference (~360 lines) with field definitions, WARNING
  reference table, annotations reference, troubleshooting quick
  reference, and available profiles.
- Rename tests/validate_tuning_report_reference.py to
  tests/validate_tuning_report_docs.py and extend validation to
  cover both documents (9 checks total: guide checks 1-6,
  reference checks 7-9).
- Update CI pipeline, source comments, config reference doc, and
  JSON schema description to reflect new file paths and wording.
